refactor commands module
