--- src/commands.rs
@@ -1,5 +1,0 @@
-pub mod dupes;
-pub mod join;
-pub mod list;
-pub mod rebuild;
-pub mod rename;
